Convert from Less CSS preprocessor to Sass.
Sass is implemented in Ruby and is supported natively in Jekyll.
Using Less required a custom plugin, the addition of the Less code in
the repository, and the availability of a JavaScript engine when
building.

Additionally this commit adds the Liberation Serif font and places it
in the default Bootstrap serif font stack.