This is a really large rewrite and revamp of some of our content.
From various sources I've heard that our content was hard to follow
or access, often our content is good but buried deep in a number
of links.

This helps to fix that by pushing the "getting started" to the
front page, it improves the getting started to be a "main hub"
to link to a set of common and important how to's, we update
users and groups with the 1.4.x dsidm commands, and I update
some other minor details.

Author: William Brown <william@blackhats.net.au>