Update development deployment to be inline with prod deployment
Rather than use an entirely different Linux distribution and deployment
strategy, this commit makes the development Dockerfile derive from the
production container image. It then installs the gems and defines an
command to render the site.  Instructions in the README then outline how
to run the container such that a user can work on site documents locally
while having the site continuously render in the container.